Small Always Encrypted Updates
We do not verify the validity of the keys or certificates used for generating the column master key.
When using Azure Key Vault as the key store, multi-tenant customer managed key rotation is not supported. Ensure that the new customer managed key is in the same tenant as the existing one.